複数のcontrollerで同じメソッドが存在していたのでhelper化
